type

+++ main.cpp
@@ -84,7 +84,7 @@ std::vector<std::vector<T>> transposed(const std::vector<std::vector<T>>& v) {
 template <typename T>
 std::vector<std::vector<T>> clockwise_rotated(
     const std::vector<std::vector<T>>& v) {
-    size_t n = v.size(), m = v.front().size();
+    int n = v.size(), m = v.front().size();
     std::vector<std::vector<T>> ret(m, std::vector<T>(n));
 
     for (int i = 0; i < m; i++) {
@@ -99,7 +99,7 @@ std::vector<std::vector<T>> clockwise_rotated(
 template <typename T>
 std::vector<std::vector<T>> counterclockwise_rotated(
     const std::vector<std::vector<T>>& v) {
-    size_t n = v.size(), m = v.front().size();
+    int n = v.size(), m = v.front().size();
     std::vector<std::vector<T>> ret(m, std::vector<T>(n));
 
     for (int i = 0; i < m; i++) {
